# openapi.yaml
openapi: 3.0.3
info:
  title: Paytm Wallet P2P API
  version: 1.0.0
  description: Concurrent peer-to-peer wallet API with Redis idempotency and PostgreSQL row-level locking.
servers:
  - url: http://localhost:4000
    description: Local API
  - url: http://localhost/api
    description: Nginx proxied API
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a new user and wallet
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: Milan
        email:
          type: string
          format: email
          example: milan@example.com
        password:
          type: string
          minLength: 8
          example: password123
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: milan@example.com
        password:
          type: string
          example: password123
    AuthResponse:
      type: object
      required: [token, user, wallet]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        wallet:
          $ref: "#/components/schemas/Wallet"
    User:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
    Wallet:
      type: object
      required: [id, userId, balancePaise, balanceDisplay, currency]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        balancePaise:
          type: integer
          example: 250000
        balanceDisplay:
          type: string
          example: INR 2,500.00
        currency:
          type: string
          example: INR
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Insufficient balance